Wiki macros can produce content for "oneliners"
Up to now, when formatting with OneLineFormatter or OutlineFormatter,
the macros were not expanded and were replaced by a `[[macro(...)]]`
indicator. Now macros can implement an optional `is_inline(content)`
method and if it returns `True`, the content will be produced and
included instead (the content should be some inline XHTML, of course).

This is now implemented for the span macro, for the html macro when
the toplevel element is an inline element, and for some formats of the
TicketQuery macro.

git-svn-id: http://svn.edgewall.org/repos/trac/trunk@10905 af82e41b-90c4-0310-8c96-b1721e28e2e2